Move `terminal` module under `tui` hierarchy
